pyproject.toml
[project]
name = "livekit-server-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "livekit-agents[openai,silero,turn-detector,deepgram,elevenlabs]~=1.1.4",
    "livekit-plugins-noise-cancellation~=0.2.4",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.0",
    "zep-cloud>=2.16.0",
]
